--- tests/test_datasets.py
@@ -1,26 +1,0 @@
-from bio_vae import datasets
-import pytest
-
-
-@pytest.mark.parametrize(
-    "dataset",
-    [
-        datasets.BBBC010,
-    ],
-)
-def test_download(dataset):
-    data = dataset()
-    assert len(data) > 0
-    assert data[0] is not None
-
-
-@pytest.mark.parametrize(
-    "dataset",
-    [
-        "BBBC010",
-    ],
-)
-def test_download(dataset):
-    data = datasets.BroadDataset(dataset)
-    assert len(data) > 0
-    assert data[0] is not None
